MAJ 23/05

+++ src/Controller/CommentController.php
@@ -39,11 +39,20 @@ class CommentController extends AbstractController
     #[Route('/comment/{id}/delete', name: 'comment_delete', methods: ['POST'])]
     public function deleteComment(Request $request, Comment $comment): Response
     {
-        $entityManager = $this->entityManager;
-        $entityManager->remove($comment);
-        $entityManager->flush();
+        // Récupérer le profil de l'utilisateur actuellement connecté
+        $currentUserProfile = $this->getUser()->getProfile();
 
-        $this->addFlash('success', 'Le commentaire a été supprimé avec succès.');
+        // Vérifier si le profil de l'utilisateur correspond au profil du commentaire
+        if ($currentUserProfile === $comment->getIdProfile()) {
+            $entityManager = $this->entityManager;
+            $entityManager->remove($comment);
+            $entityManager->flush();
+
+            $this->addFlash('success', 'Le commentaire a été supprimé avec succès.');
+        } else {
+            // Si l'utilisateur n'est pas autorisé à supprimer le commentaire, afficher un message d'erreur
+            $this->addFlash('danger', 'Vous ne pouvez pas supprimer ce commentaire.');
+        }
 
         return $this->redirectToRoute('app_tuto_video');
     }
